Fix polling gamepad
